restructured to split test.ts into multiple files
