+++ plugin-testing-framework/script/index.ts
@@ -1,6 +1,8 @@
 import * as Platform from "./platform";
 import * as PluginTestingFramework from "./test";
-import { ProjectManager } from "./projectManager";
+import { ProjectManager, setupTestRunScenario, setupUpdateScenario } from "./projectManager";
 import * as ServerUtil from "./serverUtil";
+import { TestBuilder } from "./testBuilder";
+import * as TestConfig from "./testConfig";
 import { TestUtil } from "./testUtil";
-export { Platform, PluginTestingFramework, ProjectManager, ServerUtil, TestUtil };
+export { Platform, PluginTestingFramework, ProjectManager, setupTestRunScenario, setupUpdateScenario, ServerUtil, TestBuilder, TestConfig, TestUtil };
